requirements.txt
Django==4.2.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pillow==9.5.0
celery==5.3.0
